# 11% npm install speed increase: add caching to fromUrl