@iarna 2017-06-26
The unbounded nature of this cache makes me a bit nervous, but I'm ok taking this until and if memory profiling points to change.