@mikesherov 2017-06-26
Thanks, I had a similar concern, but believe it or not, it was only slightly more memory than without the cache, presumably because of lots of memory thrashing due to unbounded object creation in npm proper.

If it ends up being a bigger deal, we can jam `lru-cache` on it, as you're suggesting.

Thanks for merging!